requirements.txt
numpy
pandas
gensim
matplotlib
mecab-python
tqdm